classify logical lock into a separte file for later pkg move
