+++ pkg/planner/core/BUILD.bazel
@@ -24,6 +24,7 @@ go_library(
         "logical_aggregation.go",
         "logical_initialize.go",
         "logical_limit.go",
+        "logical_lock.go",
         "logical_plan_builder.go",
         "logical_plans.go",
         "logical_sort.go",
